backend/requirements.txt
flask==2.2.3
flask-sqlalchemy==3.0.3
flask-migrate==4.0.4
flask-cors==3.0.10
flask-jwt-extended==4.5.2
mysqlclient==2.1.1
marshmallow==3.19.0
python-dotenv==1.0.0
werkzeug==2.2.3
requests==2.28.2
bcrypt==4.0.1
pillow==9.4.0